refact: move TngState to state.rs
