+++ src/control_interface/mod.rs
@@ -1,9 +1,6 @@
 use std::sync::Arc;
 
-use crate::{
-    config::control_interface::ControlInterfaceArgs,
-    tunnel::{RegistedService, TngState},
-};
+use crate::{config::control_interface::ControlInterfaceArgs, tunnel::{state::TngState, RegistedService}};
 use anyhow::{bail, Result};
 use async_trait::async_trait;
 use restful::RestfulControlInterface;
